refactor: TaxRecieptValidationService 관련 인터페이스 제거
