+++ src/main/java/com/seoulmilk/receipt/presentation/TaxReceiptValidationController.java
@@ -1,7 +1,7 @@
 package com.seoulmilk.receipt.presentation;
 
 import com.seoulmilk.core.presentation.RestResponse;
-import com.seoulmilk.receipt.application.TaxReceiptValidationServiceImpl;
+import com.seoulmilk.receipt.application.TaxReceiptValidationService;
 import com.seoulmilk.receipt.presentation.dto.request.TaxReceiptValidationRequest;
 import com.seoulmilk.receipt.presentation.swagger.TaxReceiptValidateSwagger;
 import io.codef.api.dto.EasyCodefResponse;
@@ -17,7 +17,7 @@ import java.util.List;
 @RequiredArgsConstructor
 @Log4j2
 public class TaxReceiptValidationController implements TaxReceiptValidateSwagger {
-    private final TaxReceiptValidationServiceImpl taxReceiptValidationService;
+    private final TaxReceiptValidationService taxReceiptValidationService;
 
     @Override
     @PostMapping("/validation")
